requirements.txt
Box2D==2.3b0
cycler==0.10.0
Cython==0.24
descartes==1.0.2
matplotlib==1.5.1
nose==1.3.7
numpy==1.11.1
psutil==4.3.0
pyparsing==2.1.5
python-dateutil==2.5.3
pytz==2016.4
PyYAML==3.11
scipy==0.17.1
Shapely==1.5.16
six==1.10.0
triangle==20160203
